research_notebooks/bitcoinenaitor/requirements.txt
paho-mqtt==1.6.1
pandas==1.5.3
pandas-ta==0.3.14b0
joblib==1.2.0
scikit-learn==1.2.2
xgboost==1.7.5 